+++ SftpLibsshAdapter/SFTPClient.cpp
@@ -6,7 +6,7 @@ namespace systelab { namespace sftp {
 
 		std::unique_ptr<IConnection>
 			SFTPClient::connect(const std::string& ip,
-								unsigned int port,
+								unsigned short port,
 								const std::string& username,
 								const std::string& pubKey,
 								const std::string& privKey,
