+++ SftpLibsshAdapter/SFTPClient.h
@@ -13,7 +13,7 @@ namespace systelab { namespace sftp {
 		~SFTPClient() = default;
 
 		std::unique_ptr<IConnection> connect(const std::string& ip,
-											 unsigned int port,
+											 unsigned short port,
 											 const std::string& username,
 											 const std::string& pubKey,
 											 const std::string& privKey,
