+++ SftpLibsshAdapter/SFTPConnection.h
@@ -16,7 +16,7 @@ namespace systelab { namespace sftp {
 		~SFTPConnection();
 
 		void connect(const std::string& ip,
-					 unsigned int port,
+					 unsigned short port,
 					 const std::string& username,
 					 const std::string& pubKey,
 					 const std::string& privKey,
